Resolves the real IR-version and opset-version collision this branch
had with main: IR12/13 were already shipped (FLOAT8E8M0, then
UINT2/INT2), and QuantizeLinear/DequantizeLinear were already bumped
to opset 25 for UINT2/INT2. This retargets FP6 to the next free slots:

- FLOAT6E2M3/FLOAT6E3M2 added as IR_VERSION 14 (enum values 27/28,
  after UINT2/INT2's 25/26) across all 5 proto files, regenerated from
  onnx.in.proto via gen_proto.py rather than hand-edited.
- QuantizeLinear/DequantizeLinear given a new opset-29 schema with
  FLOAT6E2M3/FLOAT6E3M2 added to their type constraints; the real,
  already-released opset-25 schema (with UINT2/INT2, block_size,
  precision attr, etc.) moved unmodified to old.cc as historical.
- OpSet_Onnx_ver29 registered in operator_sets.h; max ai.onnx version
  bumped 28->29 in schema.h's DomainToVersionRange.
- Added Float6E2M3/Float6E3M2 constants to the types:: helper library.

Scope deliberately excludes: bumping the other ~16 type-generic ops
(Cast, Reshape, etc.) to a new opset, and version-converter adapters
for 28<->29 -- left as follow-up, matching the original PR's own
(narrower) stated scope.

All other conflicts (parser.cc/h switch/map entries, helper.py and
numpy_helper.py packing branches, op_cast.py's Cast_25 class,
_mapping.py dtype map, docs tables) were independent additions from
both sides and were kept side by side rather than requiring a design
decision. Removed a duplicate TestFP6Reference class that had been
accidentally pasted into reference_evaluator_test.py (a byte-for-byte
copy of the new fp6_test.py), matching @justinchuby's review comment
flagging it for removal.

+++ onnx/version_converter/adapters/pad_10_11.h
@@ -1,8 +1,6 @@
 // Copyright (c) ONNX Project Contributors
-
-/*
- * SPDX-License-Identifier: Apache-2.0
- */
+//
+// SPDX-License-Identifier: Apache-2.0
 
 // Adapter for Pad in default domain from version 10 to 11
 
@@ -28,7 +26,7 @@ class Pad_10_11 final : public Adapter {
     for (int64_t shape : node->is(kpads)) {
       data_pads.emplace_back(shape);
     }
-    t_pads.sizes() = std::vector<int64_t>{(int64_t)data_pads.size()};
+    t_pads.sizes() = std::vector<int64_t>{static_cast<int64_t>(data_pads.size())};
     Value* v_pads = graph->addInitializerAndCreateValue(t_pads);
     node->addInput(v_pads);
     node->removeAttribute(kpads);
@@ -39,7 +37,7 @@ class Pad_10_11 final : public Adapter {
       Tensor t_value;
       t_value.elem_type() = TensorProto_DataType_FLOAT;
       auto& data_value = t_value.floats();
-      data_value.emplace_back(node->f(kvalue));
+      data_value.emplace_back(static_cast<float>(node->f(kvalue)));
       Node* constant = graph->create(kConstant);
       constant->insertBefore(node);
       constant->t_(kvalue, t_value);
